digital: reworked constellation soft decoder.
In packetized mode, the 'info' message input port triggers the block to process the input data and tells it how to handle the payload location and length. Can also set the constellation and other features. In this mode, no data goes out of the output streaming port and PDUs are instead emitted from the 'pdu' message port.