digital: wip: fix payload length calculation
The state machine in STATE_HAVE_HEADER must be able to count how many
symbols to skip, which is a function of the payload length (in bytes)
and the bits per symbol (bps). Previously this was only being done
correctly for the soft symbol parser; now it works for hard symbol
parser as well.