digital: fixing PFB clock sync block handling of tags.
Fixes handling of time_est tag information and how to convert to the
correct phase arm of the filterbank.

Stopped using enable_update_rate; instead, handle tags within work.

Because the relative rate can change within this block's work
function, applying the same update to the tag offets after work does
not work. Instead, we use TPP_DONT and handle the placement of the
tags on the output buffer ourselves.