digital: updates to packet_formatter classes and parsing blocks.
Emits PMT_F message when header parsing fails.

When we exit the loop prematurely, return false and the number of
items we've processed so far. This tells the parser block to emit a
PMT_F message that should be used to reset any other blocks' states
(like the header/payload demux block).

Adding loggers to packet_formatter class for easier debugging.