digital: Dealing with scaling issues with corr_est.
The correlation estimator detector assumes the input is scaled to near
+/-1. This is not a guaranteed condition and AGCs can be slow to
handle the bursts well.

Instead, this plays with a threshold based on the second moment (more
or less) of the recieved samples. Anything too far outside this range
is detected as a (potential) peak.

The second thing that this commit does is add a new "amp_est"
tag. This is the reciprocal of the amplitude of the sample when the
correlation peak is detected. It is meant so that a downstream block
can use this tag to rescale all samples by the value of this tag.