digital: adding more support for packet formatters.
A formatter class that uses a CRC8 check on the header. This is a
revision of the packet_header_default but based on the
packet_formatter_base class.

Also added a TSB version of the packet_format class that takes
packed-bytes in, packed-bytes out.

Added example for using/testing new formatters.